+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added kind `formulaBlock` to `DocumentParagraph`.
 - Added property `common_name` to `DocumentKeyValuePair`.
 - Added property `code` to `CurrencyValue`.
+- Added properties `unit`, `city_district`, `state_district`, `suburb`, `house`, and `level` to `AddressValue`.
 
 ### Breaking Changes